+++ scm-ui/ui-webapp/src/repos/branches/containers/CreateBranch.tsx
@@ -119,7 +119,7 @@ class CreateBranch extends React.Component<Props> {
 const mapDispatchToProps = (dispatch: any) => {
   return {
     fetchBranches: (repository: Repository) => {
-      dispatch(fetchBranches(repository));
+      dispatch(fetchBranches(repository, false));
     },
     createBranch: (
       createLink: string,